feat: Add getUserByCourseId function
This commit adds a new function, getUserByCourseId, to the user service. This function retrieves user data based on the course ID. It is used in multiple views, including CouseDetailView, CourseManagementView, Attendance/MappingViewVersion2, Attendance/MappingForStudent, and EnrolmentManagementView.

The function makes an HTTP GET request to the `/users/course/{courseId}` endpoint to fetch the user data. It is called in the onMounted lifecycle hook in the respective views.

This change improves the functionality and data retrieval capabilities of the application.